diagram refactoring
