+++ packages/diagram/src/module/diagrams/baseDiagramModule.ts
@@ -1,4 +1,7 @@
-import { InterpreterModule, assign, id } from "@hylimo/core";
+import { InterpreterModule } from "@hylimo/core";
+import { createDiagramModule } from "./generateDiagramModule.js";
+import { defaultStylesModule } from "./content/common/defaultStyles.js";
+import { elementModule } from "./content/common/element.js";
 
 /**
  * Module for base diagrams
@@ -7,5 +10,5 @@ export const baseDiagramModule = InterpreterModule.create(
     "base-diagram",
     ["diagram", "dsl"],
     [],
-    [assign("diagram", id("generateDiagramEnvironment").call())]
+    createDiagramModule("diagram", [defaultStylesModule, elementModule])
 );
