+++ packages/diagram/src/module/diagrams/content/uml/associations.ts
@@ -1,0 +1,44 @@
+import { InterpreterModule, parse } from "@hylimo/core";
+
+/**
+ * Module providing association operators
+ */
+export const associationsModule = InterpreterModule.create(
+    "uml/associations",
+    ["common/defaultMarkers"],
+    [],
+    [
+        ...parse(
+            `
+                this.create = scope.internal.createConnectionOperator
+                scope.-- = create()
+                scope.--> = create(
+                    endMarkerFactory = scope.defaultMarkers.arrow
+                )
+                scope.<-- = create(
+                    startMarkerFactory = scope.defaultMarkers.arrow
+                )
+                scope.<--> = create(
+                    startMarkerFactory = scope.defaultMarkers.arrow,
+                    endMarkerFactory = scope.defaultMarkers.arrow
+                )
+                scope.set("..", create(
+                    class = list("dashed-connection")
+                ))
+                scope.set("..>", create(
+                    endMarkerFactory = scope.defaultMarkers.arrow,
+                    class = list("dashed-connection")
+                ))
+                scope.set("<..", create(
+                    startMarkerFactory = scope.defaultMarkers.arrow,
+                    class = list("dashed-connection")
+                ))
+                scope.set("<..>", create(
+                    startMarkerFactory = scope.defaultMarkers.arrow,
+                    endMarkerFactory = scope.defaultMarkers.arrow,
+                    class = list("dashed-connection")
+                ))
+            `
+        )
+    ]
+);
