+++ packages/diagram/src/module/diagrams/content/uml/classifier/ports.ts
@@ -1,0 +1,11 @@
+import { InterpreterModule } from "@hylimo/core";
+
+/**
+ * Module providing helper function to create ports for a classifier
+ */
+export const portsModule = InterpreterModule.create(
+    "uml/classifier/ports",
+    [],
+    [],
+    []
+)
