+++ packages/diagram/src/module/diagrams/content/uml/classifier/providesAndRequires.ts
@@ -1,0 +1,11 @@
+import { InterpreterModule } from "@hylimo/core";
+
+/**
+ * Module providing helper function to create provided and required interfaces for a classifier
+ */
+export const providesAndRequiresModule = InterpreterModule.create(
+    "uml/classifier/providesAndRequires",
+    [],
+    [],
+    []
+)
