+++ packages/diagram/src/module/diagrams/content/uml/component.ts
@@ -1,0 +1,11 @@
+import { InterpreterModule } from "@hylimo/core";
+
+/**
+ * Module providing the component element
+ */
+export const componentModule = InterpreterModule.create(
+    "uml/component",
+    [],
+    [],
+    []
+)
